requirements.txt
fastapi
Hypercorn
asyncio
jinja2
pytz
python-dateutil
aiofiles
httpx[http2]
google-cloud-aiplatform
flask
google-cloud-discoveryengine
langchain
google-cloud-dialogflow-cx
google-api-core
grpcio
googleapis-common-protos
uuid
Pillow
pandas
google-cloud-secret-manager